JUS-105 UserDb 관련 코드를 UserMgr로 정리 * userManager.js 파일 추가 * google 작업 완료 : UserDb 접근 코드 제거 -> userMgr로 변경
